+++ opencontainers/digest/algorithm.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2020 Vanessa Sochat.
+# Copyright (C) 2019-2021 Vanessa Sochat.
 
 # This Source Code Form is subject to the terms of the
 # Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
@@ -19,7 +19,9 @@ import io
 
 
 class Algorithm(StrStruct):
-    """Algorithm identifies and implementation of a digester by an identifier.
+    """
+    Algorithm identifies and implementation of a digester by an identifier.
+
     Note the that this defines both the hash algorithm used and the string
     encoding.
     """
@@ -29,7 +31,9 @@ class Algorithm(StrStruct):
         super().__init__(value)
 
     def available(self):
-        """Available returns true if the digest type is available for use.
+        """
+        Available returns true if the digest type is available for use.
+
         If this returns false, Digester and Hash will return None.
         we are flexible to allow the user to also provide a full digest
         """
@@ -45,23 +49,30 @@ class Algorithm(StrStruct):
         return algorithm in algorithms
 
     def digester(self):
-        """Digester returns a new digester for the specified algorithm. If the algorithm
-        does not have a digester implementation, nil will be returned. This can be
-        checked by calling Available before calling Digester. Note that
-        the GoLang implementation also had a Hash() function that (seemed to)
-        return the same, and instead I'm going to return the same hashlib new.
+        """
+        Digester returns a new digester for the specified algorithm.
+
+        If the algorithm does not have a digester implementation, nil will be
+        returned. This can be checked by calling Available before calling
+        Digester. Note that the GoLang implementation also had a Hash() function
+        that (seemed to) return the same, and instead I'm going to return the
+        same hashlib new.
         """
         return digester(self, self.hash())
 
     def hash(self):
-        """Hash returns a new hash as used by the algorithm."""
+        """
+        Hash returns a new hash as used by the algorithm.
+        """
         if not self.available():
             return None
         return hashlib.new(self._algorithm)
 
     def validate(self, encoded):
-        """Validate validates the encoded portion string. This means
-        ensuring that the algorithm is available, checking it's length,
+        """
+        Validate validates the encoded portion string.
+
+        This means ensuring that the algorithm is available, checking it's length,
         and the characters provided.
         """
         if not self.available():
@@ -79,7 +90,9 @@ class Algorithm(StrStruct):
         return True
 
     def size(self):
-        """Size returns number of bytes returned by the hash."""
+        """
+        Size returns number of bytes returned by the hash.
+        """
         if not self.available():
             return 0
         hashy = hashlib.new(self._algorithm)
@@ -88,7 +101,9 @@ class Algorithm(StrStruct):
         return hashy.digest_size
 
     def set(self, value):
-        """Set implemented to allow use of Algorithm as a command line flag.
+        """
+        Set implemented to allow use of Algorithm as a command line flag.
+
         This isn't useful, as we could already call load (but this will
         mirror GoLang.
         """
@@ -97,8 +112,10 @@ class Algorithm(StrStruct):
             raise ErrDigestUnsupported()
 
     def encode(self, content):
-        """Encode encodes the raw bytes of a digest, typically from a hash.Hash, into
-        the encoded portion of the digest.
+        """
+        Encode encodes the raw bytes of a digest, typically from a hash.
+
+        into the encoded portion of the digest.
         """
         # Currently, all algorithms use a hex encoding. When we
         # add support for back registration, we can modify this accordingly.
@@ -108,7 +125,9 @@ class Algorithm(StrStruct):
         return content.hex()
 
     def fromReader(self, ioReader):
-        """FromReader returns the digest of the reader using the algorithm.
+        """
+        FromReader returns the digest of the reader using the algorithm.
+
         the input must be type bytes, usually from io.BytesIO.read().
         This function likely isn't needed, but is provided to mirror
         the GoLang implementation.
@@ -124,7 +143,9 @@ class Algorithm(StrStruct):
         return digester.digest()
 
     def fromString(self, content):
-        """FromString digests the string input and returns a Digest.
+        """
+        FromString digests the string input and returns a Digest.
+
         TODO not sure what this is intended for.
         """
         if not isinstance(content, str):