+++ opencontainers/image/v1/config.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2020 Vanessa Sochat.
+# Copyright (C) 2019-2021 Vanessa Sochat.
 
 # This Source Code Form is subject to the terms of the
 # Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
@@ -11,7 +11,10 @@ from datetime import datetime
 
 
 class ImageConfig(Struct):
-    """ImageConfig defines the execution parameters which should be used as a
+    """
+    An ImageConfig structure.
+
+    ImageConfig defines the execution parameters which should be used as a
     base when running a container using an image.
     """
 
@@ -71,7 +74,9 @@ class ImageConfig(Struct):
 
 
 class RootFS(Struct):
-    """RootFS describes a layer content addresses"""
+    """
+    RootFS describes a layer content addresses
+    """
 
     def __init__(self, rootfs_type=None, diff_ids=None):
         super().__init__()
@@ -89,7 +94,9 @@ class RootFS(Struct):
 
 
 class History(Struct):
-    """History describes the history of a layer."""
+    """
+    History describes the history of a layer.
+    """
 
     def __init__(
         self, created=None, created_by=None, author=None, comment=None, empty_layer=None
@@ -120,7 +127,10 @@ class History(Struct):
 
 
 class Image(Struct):
-    """Image is the JSON structure which describes some basic information about
+    """
+    An Image Structure
+
+    Image is the JSON structure which describes some basic information about
     the image. This provides the `application/vnd.oci.image.config.v1+json`
     mediatype when marshalled to JSON.
     """